requirements.txt
Django==5.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==11.3.0
python-decouple==3.8
requests==2.31.0
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.0
